separate inserting data in PinInput component logic into custom hook
